Fixed critical permission bug

+++ commands/mod/purge.js
@@ -18,6 +18,7 @@ module.exports = class PurgeCommand extends Commando.Command {
   }
 
   run(message, args) {
+    if (!modRole[message.guild.id]) return message.reply("There are no roles set up for this comamnd to run");
     if (message.member.roles.some(r => modRole[message.guild.id].modroles.includes(r.id) || message.author.id === message.guild.ownerID)) {
       message.channel.bulkDelete(args.number);
       return message.delete();
